Fixed controller and system tests in the engine
Getting the routes to work correctly within the engine was a bit
difficult. The controller tests just required the routes to be included
but the system tests were a bit more difficult, so I looked up the
parent class and overrode a particular method that provided the url
completions.

Gems from the main app were required to pass the tests while running
them from the engine's directory. Some of the gem probably could have
been removed from both files so there wasn't so much duplication.